+++ zcash_client_backend/src/proposal.rs
@@ -11,6 +11,7 @@ use zcash_protocol::{consensus::BlockHeight, value::Zatoshis, PoolType, Shielded
 use zip321::TransactionRequest;
 
 use crate::{
+    data_api::wallet::TargetHeight,
     fees::TransactionBalance,
     wallet::{Note, ReceivedNote, WalletTransparentOutput},
 };
@@ -166,7 +167,7 @@ impl<NoteRef> ShieldedInputs<NoteRef> {
 #[derive(Clone, PartialEq, Eq)]
 pub struct Proposal<FeeRuleT, NoteRef> {
     fee_rule: FeeRuleT,
-    min_target_height: BlockHeight,
+    min_target_height: TargetHeight,
     steps: NonEmpty<Step<NoteRef>>,
 }
 
@@ -183,7 +184,7 @@ impl<FeeRuleT, NoteRef> Proposal<FeeRuleT, NoteRef> {
     /// * `steps`: A vector of steps that make up the proposal.
     pub fn multi_step(
         fee_rule: FeeRuleT,
-        min_target_height: BlockHeight,
+        min_target_height: TargetHeight,
         steps: NonEmpty<Step<NoteRef>>,
     ) -> Result<Self, ProposalError> {
         let mut consumed_chain_inputs: BTreeSet<(PoolType, TxId, u32)> = BTreeSet::new();
@@ -273,7 +274,7 @@ impl<FeeRuleT, NoteRef> Proposal<FeeRuleT, NoteRef> {
         shielded_inputs: Option<ShieldedInputs<NoteRef>>,
         balance: TransactionBalance,
         fee_rule: FeeRuleT,
-        min_target_height: BlockHeight,
+        min_target_height: TargetHeight,
         is_shielding: bool,
     ) -> Result<Self, ProposalError> {
         Ok(Self {
@@ -301,7 +302,7 @@ impl<FeeRuleT, NoteRef> Proposal<FeeRuleT, NoteRef> {
     ///
     /// The chain must contain at least this many blocks in order for the proposal to
     /// be executed.
-    pub fn min_target_height(&self) -> BlockHeight {
+    pub fn min_target_height(&self) -> TargetHeight {
         self.min_target_height
     }