+++ zcash_client_sqlite/src/testing/db.rs
@@ -12,12 +12,13 @@ use tempfile::NamedTempFile;
 use rusqlite::{self};
 use secrecy::SecretVec;
 use shardtree::{error::ShardTreeError, ShardTree};
+
 use zcash_client_backend::{
     data_api::{
         chain::{ChainState, CommitmentTreeRoot},
         scanning::ScanRange,
         testing::{DataStoreFactory, Reset, TestState},
-        wallet::ConfirmationsPolicy,
+        wallet::{ConfirmationsPolicy, TargetHeight},
         TargetValue, *,
     },
     wallet::{Note, NoteId, ReceivedNote, WalletTransparentOutput},